Re-code genotypes according to the supplied genotype-phenotype map

Given a genotype-phenotype (GP) map recodes genotypes in G. The GP map is a matrix with each row containing new coding for one marker in G: new codes for genotypes 'AA', 'aA' and 'aa' (in G encoded as 0, 1 and 2, respectively).

recode_G(G, GP_map)

Arguments

G
genotypes matrix

GP_map
genotype-phenotype map

Value

weights matrix

Details

If only a vector GP map is provided, same re-coding is assumed for each and every locus in G.

Author

Marcin Kierczak <marcin.kierczak@scilifelab.se>